annotators/speech_function_predictor/requirements.txt
Flask==2.2.3
gunicorn==19.9.0
sentry_sdk==0.13.3
requests>=2.25.1
jinja2<=3.0.3
Werkzeug==2.2.2
allure-pytest==2.13.2